% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_effects.R
\name{sem_eff_tab_v2}
\alias{sem_eff_tab_v2}
\title{Process SEM effects output for OE,MMI, EPT relabel variable names to make into a table}
\usage{
sem_eff_tab_v2(sem_output)
}
\arguments{
\item{sem_output}{SEM estimated effects output saved as .csv file}
}
\value{
Processed dataframe that selected direct, indirect, and total effects
}
\description{
Process SEM effects output for OE,MMI, EPT relabel variable names to make into a table
}
\examples{
xer_oe_m2_full<- sem_eff_tab("model2") # name dataframe loaded with raw SEM effects output
}
