% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_r2.R
\name{sem_r2_tab_v2}
\alias{sem_r2_tab_v2}
\title{Process SEM R2 output for OE,MMI, EPT relabel variable names to make into a table}
\usage{
sem_r2_tab_v2(R2_output)
}
\arguments{
\item{R2_output}{SEM estimated R2 for endogenous variables saved as .csv file}
}
\value{
Processed dataframe that R2 for endogenous variables in the model
}
\description{
Process SEM R2 output for OE,MMI, EPT relabel variable names to make into a table
}
\examples{
west_oe_m1_full<- r2_tab("model1") # name R2 dataframe processed from SEM output

}
