% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snTest.R
\name{snTest}
\alias{snTest}
\title{Signal to Noise test SCRIPT FROM KAREN BLOCKSOM EMAIL 2/10/2023}
\usage{
snTest(dfIn, idVars.samp, idVars.site, year = "YEAR")
}
\arguments{
\item{dfIn}{This data frame is in wide format with one row per sample and assumed to contain only numeric metrics.
It should also contain all visits to each site, and at least a subset of sites must have multiple visits.}

\item{idVars.samp}{A character vector containing variables that identify individual samples. Use UID for idVars.samp
- should be unique among all samples in the data}

\item{idVars.site}{A string containing variable name that identifies sites.
This cannot be the same as or a subset of variables in idVars.samp. Use UNIQUE_ID for idVars.site - if using multiple cycles of data}

\item{year}{A string containing name of Year variable if sites are revisited across years (as well as within year),
default is 'YEAR'. Set to NULL if no samples across years.}
}
\value{
dataframe of signal to noise ratios
}
\description{
Signal to Noise test SCRIPT FROM KAREN BLOCKSOM EMAIL 2/10/2023
}
